Backend/requirements.txt
Django==5.2.8
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.1
Pillow==10.0.0
numpy==1.24.3
torch==2.0.1
transformers==4.30.2
Pillow==10.0.0
django-filter==23.2
huggingface-hub==0.16.4
open-clip-torch==2.24.0